refactor(entities): se cambio el tipo de dato del id a long (se me olvido incluirlo en el branch "Int-To-Long")

+++ app/src/main/java/com/cashcontrol/data/local/entities/CategoriaEntity.kt
@@ -7,7 +7,7 @@ import androidx.room.PrimaryKey
 data class CategoriaEntity(
     @PrimaryKey
     val categoriaId: Long?,
-    val usuarioId: Int?,
+    val usuarioId: Long?,
     val tipo: String,
     val descripcion: String
 )
